+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # Inbound changes not assigned to version # yet
 - Use `parallel_sync` Python package for file transfers
 - Add close() in update_status to force use of a fresh connection on the next update
+- chmod's the output files using concurrent.futures
 
 ## 1.1.2 Summary [05/15/2023]